requirements.txt
Django>=4.2
Pillow>=10.0
django-unfold>=0.79.0
